Move ballista information to separate component
