+++ .prettierrc
@@ -1,7 +1,7 @@
 {
   "arrowParens": "avoid",
   "bracketSpacing": true,
-  "printWidth": 80,
+  "printWidth": 160,
   "semi": true,
   "singleQuote": true,
   "tabWidth": 2
